Splitting directive utils from the rest.
